glassbox_ai/requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Image processing
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0

# Web interface
streamlit>=1.35.0

# Utilities
PyYAML>=6.0
requests>=2.31.0